トリミングできるようにした
