+++ native/android/app/src/main/res/navigation/navigation.xml
@@ -3,7 +3,7 @@
   xmlns:app="http://schemas.android.com/apk/res-auto"
   xmlns:tools="http://schemas.android.com/tools"
   android:id="@+id/navigation"
-  app:startDestination="@id/groupListFragment">
+  app:startDestination="@id/login_fragment">
 
   <fragment
     android:id="@+id/login_fragment"
